Use container rdfs:member item instead of item rdfs:member container
rdfs:member is an ambiguous name for a property; it could either mean
that the subject is a member of the object, or that the object is a
member of the subject. The former would more clearly be named
"memberOf" and is what has been suggested in the documentation before
this change. The latter would more clearly be named "hasMember" and
appears to be the correct interpretation of the property.

For a motivation why the containing collection should be the subject
and not the object, see the following references:

https://www.w3.org/TR/rdf-schema/#ch_containermembershipproperty
https://www.w3.org/TR/rdf-schema/#ch_sumproperties