Add a selector type for edpoprec:Record fields
The existing PredicateSelector does not fit edpoprec:Records very
well. The field as a whole can be identified by a property, but the
potentially multiple field values cannot be easily distinguished
because they are blank records. The originalText property of those
blank records could be used to identify them instead. While it might
be possible to nest this information in the edpopcol:object, this
would complicate application logic. Hence, a dedicated selector for
record fields seems more appropriate.

The edpoprec: prefix is referenced for documentation purposes.